+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added support for transparently following debug links in ELF binaries
+  - Added `symbolize::Builder::set_debug_dirs` for configuring directories
+    searched for targets
 - Fixed handling of zero sized symbols in Gsym symbolization logic
 - Fixed reading of ELF section contents of sections without actual data